refactor(site): rename new-header to header
Completes the prune-and-rename that "feat: new header navbar (#868)" started.
Now that the superseded component is gone, the name is free.

new-header/ -> header/, new-header.tsx -> header.tsx, NewHeader -> Header,
NewHeaderProps -> HeaderProps. The barrel export and (marketing)/layout.tsx
follow. Header is reordered into the barrel's alphabetical position.

No behavior change; the rendered component is byte-identical.
