+++ apps/site/src/app/(unauthenticated)/[locale]/(marketing)/components/index.ts
@@ -1,6 +1,7 @@
 // Copyright © Todd Agriscience, Inc. All rights reserved.
 
 export { default as Footer } from './footer/footer';
+export { default as Header } from './header/header';
 export { default as Homepage } from './homepage/homepage';
 export {
   MARKETING_PILL_LINK_CLASSNAME,
@@ -17,6 +18,5 @@ export {
   MarketingValuesOperatingStack,
   MarketingValuesTwoColumns,
 } from './marketing-blocks';
-export { default as NewHeader } from './new-header/new-header';
 export { default as NewsHighlights } from './news-highlights/news-highlights';
 export { default as ScrollShrinkWrapper } from './scroll-shrink-wrapper/scroll-shrink-wrapper';
